Update MongoUtils.cs
Changes:

I've changed the var's, because is not good regarding the performance use var, because you know perfectly what thos functions are returning, so there is no need to let the program, to infere. And return that directly, there is no need to store the hash in a string and then return. 

About the CamelCase, is the string value was an empty string, the program will crash with an ArgumentOutOfRangeException, so a simple check will fix that part.

And at the end, only to standarized, it's cleaner to standarized all the "" to String.Empty.