Fixing method summaries to be coherent
Fixing the summary of both "GetSerializer" methods to be coherent with what they actually do. 
They not only gets a registered serializer, but they also ADDs it when it wasn't registered before (behavior I disagree BTW). And it took me a while to find out this was the problem I was facing.