Add MongoDB.Driver support for big-endian clients
The current MongoDB.Driver causes a crash when used on big-endian
systems. This error stems from the use of BitConverter function
used in method definitions across the code.

The BitConverter class does conversions according to the endianess
of the system it is operating on which causes incorrect reads of
little-endian packets sent by the server.

BitConverter can be replaced with BinaryPrimitives provided by C#
but due to their lack of support for netstandard2.0 and to maintain
backcompatibility implemented a PacketBitConverter under
MongoDB.Bson.IO namespace.