CSHARP-5778: Fix atlas-search-index-helpers-test
Turns out that the version on Evergreen is 8.0.16, while the version I have locally is 8.2.1. The exact exception message content is different between the two.

I have tested this with a patch on Evergreen.

There does not appear to be a way to specify the minor version for an Atlas deployment in Evergreen.